moved __tests__ to src directory
